Switching to web sockets and web based interface.
I wanted a way to be able to know if somebody was actively controlling the
bot. If not I want it to stand still.

Using websockets lets me hold a connection and shutdown the bot when
nobody is connected. It also makes it easy to ensure only one person /
controller is in charge while not requiring me to implementing my own binary
protocol.